Set product and product context types
