--- client/src/contexts/ProductContext.jsx
@@ -1,20 +1,0 @@
-import { createContext, useState, useEffect } from "react";
-
-export const ProductContext = createContext();
-
-const ProductProvider = ({ children }) => {
-  // products state
-  const [products, setProducts] = useState ([]);
-  // fetch products
-  useEffect(() => {
-    const fetchProducts = async () => {
-      const response = await fetch('https://fakestoreapi.com/products');
-      const data = await response.json();
-      setProducts(data);
-    }
-    fetchProducts();
-    }, []);
-  return <ProductContext.Provider value={{products}}>{children}</ProductContext.Provider>;
-};
-
-export default ProductProvider;
