+++ client/src/contexts/ProductContext.tsx
@@ -1,0 +1,36 @@
+import { createContext, useState, useEffect, type ReactNode } from "react";
+
+export type ProductType = {
+  id: number,
+  title: string,
+  price: number,
+  description: string,
+  category: string,
+  image: string,
+  rating: {
+    rate: number,
+    count: number,
+  },
+}
+export type ProductContextType = {
+  products: ProductType[],
+}
+
+export const ProductContext = createContext<ProductContextType | null>(null);
+
+const ProductProvider = ({ children }: { children: ReactNode }) => {
+  // products state
+  const [products, setProducts] = useState ([]);
+  // fetch products
+  useEffect(() => {
+    const fetchProducts = async () => {
+      const response = await fetch('https://fakestoreapi.com/products');
+      const data = await response.json();
+      setProducts(data);
+    }
+    fetchProducts();
+    }, []);
+  return <ProductContext.Provider value={{products}}>{children}</ProductContext.Provider>;
+};
+
+export default ProductProvider;
